Reorganize scripts in email_backend
