+++ email_backend/.gitignore
@@ -7,4 +7,4 @@ npm-debug*
 coverage/
 *credentials*.json
 transpiled_build/
-data_management/CSVs/
+scripts/data_management/CSVs/
