+++ email_backend/scripts/data_management/run_local_server_in_prod_mode.sh
@@ -1,0 +1,17 @@
+#!/bin/bash
+# must run this scripts in the email_backend dir
+set -e # will exit if any command has non-zero exit value
+
+npm run build
+
+source scripts/util/set_prod_env_vars.sh
+export IS_FAKE_PROD_SERVER=true
+
+source scripts/util/set_transient_secrets.sh # these trap EXIT to handle their own cleanup
+
+npm run start
+
+source scripts/util/unset_prod_env_vars.sh
+unset IS_FAKE_PROD_SERVER=true
+
+exit
